Flask-Product-CRUD-main/requirements.txt
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
PyMySQL==1.0.2
python-dotenv==0.20.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
WTForms==3.0.1
